Added Offener Posten and multiple Repositories, several fixes in test classes and Auftrag, Kunde and Kundenanfrage

+++ src/main/java/com/iu/kmi/test/KonditionTest.java
@@ -5,6 +5,8 @@ import org.junit.Test;
 
 import com.iu.kmi.entities.Kondition;
 
+import java.math.BigDecimal;
+
 import static org.junit.Assert.assertEquals;
 
 /**
@@ -19,13 +21,13 @@ public class KonditionTest {
     @Test
     public void testConstructorAndGetters() {
         // Erstellt eine neue Kondition mit den angegebenen Parametern
-        Kondition kondition = new Kondition("Standard", "Kreditkarte", "Lieferung in 5 Tagen", 10.0);
+        Kondition kondition = new Kondition("Standard", "Kreditkarte", "Lieferung in 5 Tagen", BigDecimal.valueOf(10.0));
 
         // Überprüft, ob die Getter-Methoden die erwarteten Werte zurückgeben
         assertEquals("Standard", kondition.getName());
         assertEquals("Kreditkarte", kondition.getZahlungsbedingungen());
         assertEquals("Lieferung in 5 Tagen", kondition.getLieferbedingungen());
-        assertEquals(10.0, kondition.getRabatt());
+        assertEquals(BigDecimal.valueOf(10.0), kondition.getRabatt());
     }
 
     /**
@@ -34,21 +36,21 @@ public class KonditionTest {
     @Test
     public void testSetters() {
         // Erstellt eine neue Kondition mit den angegebenen Parametern
-        Kondition kondition = new Kondition("Standard", "Kreditkarte", "Lieferung in 5 Tagen", 10.0);
+        Kondition kondition = new Kondition("Standard", "Kreditkarte", "Lieferung in 5 Tagen", BigDecimal.valueOf(10.0));
 
         // Setzt neue Werte für die Eigenschaften der Kondition
         kondition.setKonditionNr("1");
         kondition.setName("Premium");
         kondition.setZahlungsbedingungen("PayPal");
         kondition.setLieferbedingungen("Lieferung in 2 Tagen");
-        kondition.setRabatt(15.0);
+        kondition.setRabatt(BigDecimal.valueOf(15.0));
 
         // Überprüft, ob die Setter-Methoden die neuen Werte korrekt setzen
         assertEquals("1", kondition.getKonditionNr());
         assertEquals("Premium", kondition.getName());
         assertEquals("PayPal", kondition.getZahlungsbedingungen());
         assertEquals("Lieferung in 2 Tagen", kondition.getLieferbedingungen());
-        assertEquals(15.0, kondition.getRabatt());
+        assertEquals(BigDecimal.valueOf(15.0), kondition.getRabatt());
     }
 
     /**
@@ -57,7 +59,7 @@ public class KonditionTest {
     @Test
     public void testToString() {
         // Erstellt eine neue Kondition mit den angegebenen Parametern
-        Kondition kondition = new Kondition("Standard", "Kreditkarte", "Lieferung in 5 Tagen", 10.0);
+        Kondition kondition = new Kondition("Standard", "Kreditkarte", "Lieferung in 5 Tagen", BigDecimal.valueOf(10.0));
 
         // Setzt die KonditionenNr der Kondition
         kondition.setKonditionNr("1");
